feat: 정규식 구현 및 phoneNumber format 자동 수정 기능 구현
