+++ ios-contact-manager/Support/String+.swift
@@ -1,57 +1,43 @@
 import UIKit
 
-enum CustomString: CustomStringConvertible {
+extension String {
+    
+    func formatPhoneNumber() -> String {
+        var formatText = self.components(separatedBy: CharacterSet.decimalDigits.inverted).joined()
 
-    case nameText
-    case ageText
-    case phoneNumberText
-    case cellNameText
-    case wrongNameMessage
-    case wrongAgeMessage
-    case wrongPhoneNumberMessage
-    case cancelMessage
-    case detailTitleText
-    case contactListTitleText
-    case fatalErrorText
-    case RegularExpressionNamePatthenText
-    case RegularExpressionAgePatthenText
-    case RegularExpressionPhoneNumberPatthenText
-    case RegularMatchesText
-    
-    var description: String {
-        switch self {
-            
-        case .nameText:
-            return "이름"
-        case .ageText:
-            return "나이"
-        case .phoneNumberText:
-            return "연락처"
-        case .cellNameText:
-            return "contactCell"
-        case .wrongNameMessage:
-            return "입력한 이름 정보가 잘못되었습니다."
-        case .wrongAgeMessage:
-            return "입력한 나이정보가 잘못되었습니다."
-        case .wrongPhoneNumberMessage:
-            return "입력한 연락처 정보가 잘못되었습니다."
-        case .cancelMessage:
-            return "정말 종료하시겠습니까?"
-        case .detailTitleText:
-            return "새 연락처"
-        case .contactListTitleText:
-            return "연락처"
-        case .fatalErrorText:
-            return "init(coder:) has not been implemented"
-        case .RegularExpressionNamePatthenText:
-            return "^[a-zA-Z가-힣]*$"
-        case .RegularExpressionAgePatthenText:
-            return "^[0-9]{1,3}$"
-        case .RegularExpressionPhoneNumberPatthenText:
-            return "^\\d{2,3}-\\d{3,4}-\\d{4}$"
-        case .RegularMatchesText:
-            return "SELF MATCHES %@"
+        guard !formatText.isEmpty else {
+            return ""
         }
+
+        let hyphen: Character = "-"
+
+        if formatText.hasPrefix("02") {
+            switch formatText.count {
+            case 3...5:
+                formatText.insert(hyphen, at: formatText.index(formatText.startIndex, offsetBy: 2))
+            case 6...9:
+                formatText.insert(hyphen, at: formatText.index(formatText.startIndex, offsetBy: 2))
+                formatText.insert(hyphen, at: formatText.index(formatText.endIndex, offsetBy: 6 - formatText.count))
+            case 10:
+                formatText.insert(hyphen, at: formatText.index(formatText.startIndex, offsetBy: 2))
+                formatText.insert(hyphen, at: formatText.index(formatText.endIndex, offsetBy: -4))
+            default:
+                break
+            }
+        } else {
+            switch formatText.count {
+            case 4...6:
+                formatText.insert(hyphen, at: formatText.index(formatText.startIndex, offsetBy: 3))
+            case 7...10:
+                formatText.insert(hyphen, at: formatText.index(formatText.startIndex, offsetBy: 3))
+                formatText.insert(hyphen, at: formatText.index(formatText.endIndex, offsetBy: 7 - formatText.count))
+            case 11:
+                formatText.insert(hyphen, at: formatText.index(formatText.startIndex, offsetBy: 3))
+                formatText.insert(hyphen, at: formatText.index(formatText.endIndex, offsetBy: -4))
+            default:
+                break
+            }
+        }
+        return formatText
     }
-    
 }
